# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Car Accessories Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with optional filtering
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Add a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - stock
        - featured
        - reviewCount
        - tags
        - imageUrls
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
...
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - price
        - stock
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        compareAtPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: ["integer", "null"]
        featured:
          type: ["boolean", "null"]
        tags:
          type: array
          items:
            type: string
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        title:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - reviewerName
        - rating
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        reviewerName:
          type: string
        rating:
          type: integer
        title:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
      required:
        - reviewerName
        - rating
...