# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  stock: number;
  featured?: boolean;
  tags?: string[];
}