# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  reviewerName: string;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  comment?: string | null;
}