# lib/api-zod/src/generated/types/updateProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductRequest {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  stock?: number | null;
  /** @nullable */
  featured?: boolean | null;
  tags?: string[];
}